base de datos creada

+++ src/main/java/com/wine_api/wine_api/controller/WineController.java
@@ -33,7 +33,7 @@ public class WineController {
 	}
 	
 	@GetMapping("/wine/{id}")
-	public ResponseEntity<Wine> getWine(@PathVariable Long id){
+	public ResponseEntity<Wine> getWine(@PathVariable Integer id){
 		
 		Wine wine = wineService.getWineById(id);
